requirements_cuda.txt
absl-py==1.0.0
astunparse==1.6.3
bcrypt==3.1.7
blessings==1.6
blinker==1.4
cachetools==4.2.4
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.0.9
click==7.1.2
colorama==0.4.4
cryptography==3.3.2
cycler==0.11.0
Cython==0.29.25
distro==1.5.0
distro-info==1.0
easydict==1.9
fasteners==0.14.1
ffmpeg==1.4
flatbuffers==2.0
fonttools==4.28.3
freetype-py==2.2.0
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpustat==0.6.0
GPUtil==1.4.0
grpcio==1.42.0
h5py==3.6.0
httplib2==0.18.1
idna==2.10
importlib-metadata==4.8.2
imutils==0.5.4
iso8601==1.0.2
isodate==0.6.1
jeepney==0.7.1
keras==2.7.0
Keras-Preprocessing==1.1.2
keyring==23.0.1
kiwisolver==1.3.2
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
libclang==12.0.0
lockfile==0.12.2
lxml==4.7.1
m3u8==0.9.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==1.1.1
matplotlib==3.5.1
MediaInfo==0.0.9
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.9
numpy==1.21.4
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
paramiko==2.7.2
pexpect==4.8.0
Pillow==8.1.2
protobuf==3.12.4
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycocotools==2.0.3
pycountry==20.7.3
pycryptodome==3.12.0
PyGObject==3.40.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyparsing==3.0.6
pyRFC3339==1.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
pyxdg==0.27
PyYAML==5.3.1
reportlab==3.5.66
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scipy==1.7.3
SecretStorage==3.3.1
simplejson==3.17.2
six==1.16.0
ssh-import-id==5.11
streamlink==3.0.3
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
thop==0.0.31.post2005241907
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tqdm==4.62.3
typing-extensions==4.0.1
urllib3==1.26.5
wadllib==1.3.5
websocket-client==1.2.3
Werkzeug==2.0.2
wrapt==1.13.3
xdg==5
zipp==1.0.0
